refactor: rename and reorganize Search component imports for better structure
